init: set up unit tester and some test cases
The unit tester is written in Python, but it is agnostic of the language
used to implement the interpreter being tested. This is because it
invokes the implemented interpreter as a subprocess and compares its
output instead of attempting to import any implementations directly.
This makes it a central test site to drive the development of Brainfuck
interpreters in any implementation language.